Per Ross Smith II:
The would work as well:

if [[ $PACKER_BUILDER_TYPE =~ vmware ]]; then
I've standardize on using [[ per

http://google-styleguide.googlecode.com/svn/trunk/shell.xml#Test,_%5B_and_%5B%5B

Has the advantage of being backwards-compatible with the old form

+++ template/fedora/script/vmtools.sh
@@ -1,8 +1,6 @@
 #!/bin/bash -eux
 
-case "$PACKER_BUILDER_TYPE" in 
-
-vmware-iso|vmware-ovf)
+if [[ $PACKER_BUILDER_TYPE =~ vmware ]]; then
     # Uninstall fuse to fake out the vmware install so it won't try to
     # enable the VMware blocking filesystem
     yum erase -y fuse
@@ -96,9 +94,9 @@ vmware-iso|vmware-ovf)
     rm /home/vagrant/linux.iso
     umount /mnt/cdrom
     rmdir /mnt/cdrom
-    ;;
+fi
 
-virtualbox-iso|virtualbox-ovf)
+if [[ $PACKER_BUILDER_TYPE =~ virtualbox ]]; then
     echo "Installing VirtualBox guest additions"
 
     # Assume that we've installed all the prerequisites:
@@ -110,11 +108,4 @@ virtualbox-iso|virtualbox-ovf)
     sh /mnt/VBoxLinuxAdditions.run --nox11
     umount /mnt
     rm -rf /home/vagrant/VBoxGuestAdditions_${VBOX_VERSION}.iso
-    ;;
-
-*)
-    echo "Unknown Packer Builder Type >>$PACKER_BUILDER_TYPE<< selected."
-    echo "Known are virtualbox-iso|virtualbox-ovf|vmware-iso|vmware-ovf."
-    ;;
-
-esac
+fi
